+++ src/preload.js
@@ -22,6 +22,7 @@ contextBridge.exposeInMainWorld('pebble', {
   getEffectsState: (output) => ipcRenderer.invoke('effects:get-state', output),
   setEffect: (id, changes, output) => ipcRenderer.invoke('effects:set', id, changes, output),
   setEffectsMaster: (enabled, output) => ipcRenderer.invoke('effects:set-master', enabled, output),
+  applySoundMode: (id, output) => ipcRenderer.invoke('effects:apply-sound-mode', id, output),
   getEqState: (output) => ipcRenderer.invoke('eq:get-state', output),
   setEq: (changes, output) => ipcRenderer.invoke('eq:set', changes, output),
   getMicState: () => ipcRenderer.invoke('mic:get-state'),